sustainability_rag/src/processors/chart_processor.py
# src/processors/chart_processor.py

import torch
from typing import Dict, List, Union, Optional
from pathlib import Path
import logging
from PIL import Image

from imagebind import data
from imagebind.models import imagebind_model
from ..utils.device_utils import get_device

class ChartProcessor:
    def __init__(
        self,
        device: Optional[str] = None
    ):
        self.device = device or get_device()
        self.model = imagebind_model.ImageBindModel(
            pretrained=True
        ).to(self.device)
        self.logger = logging.getLogger(__name__)

    async def process_chart(
        self,
        image_path: Union[str, Path],
        metrics_schema: Dict
    ) -> Dict[str, float]:
        """
        處理圖表並提取數據
        """
        try:
            # 1. 加載和預處理圖片
            image = Image.open(image_path)
            inputs = {
                "image": data.load_and_transform_image(
                    image_path, 
                    device=self.device
                )
            }

            # 2. 生成圖片嵌入
            with torch.no_grad():
                embeddings = self.model(inputs)
                image_features = embeddings["image"]

            # 3. 分析圖表類型
            chart_type = self._identify_chart_type(image_features)

            # 4. 根據圖表類型提取數據
            if chart_type == "line":
                extracted_data = self._process_line_chart(image)
            elif chart_type == "bar":
                extracted_data = self._process_bar_chart(image)
            elif chart_type == "pie":
                extracted_data = self._process_pie_chart(image)
            else:
                self.logger.warning(f"Unsupported chart type: {chart_type}")
                return {}

            # 5. 匹配指標schema
            matched_metrics = self._match_metrics(
                extracted_data,
                metrics_schema
            )

            return matched_metrics

        except Exception as e:
            self.logger.error(f"Error processing chart: {str(e)}")
            raise

    def _identify_chart_type(
        self,
        features: torch.Tensor
    ) -> str:
        """
        識別圖表類型
        """
        # TODO: 實現圖表類型識別邏輯
        return "line"

    def _process_line_chart(
        self,
        image: Image.Image
    ) -> Dict[str, float]:
        """
        處理折線圖
        """
        # TODO: 實現折線圖數據提取邏輯
        return {}

    def _process_bar_chart(
        self,
        image: Image.Image
    ) -> Dict[str, float]:
        """
        處理柱狀圖
        """
        # TODO: 實現柱狀圖數據提取邏輯
        return {}

    def _process_pie_chart(
        self,
        image: Image.Image
    ) -> Dict[str, float]:
        """
        處理圓餅圖
        """
        # TODO: 實現圓餅圖數據提取邏輯
        return {}

    def _match_metrics(
        self,
        extracted_data: Dict[str, float],
        metrics_schema: Dict
    ) -> Dict[str, float]:
        """
        將提取的數據匹配到指標schema
        """
        matched_metrics = {}
        
        for metric_id, metric_info in metrics_schema.items():
            if metric_info["name"] in extracted_data:
                matched_metrics[metric_id] = extracted_data[metric_info["name"]]

        return matched_metrics