sustainability_rag/requirements.txt
# Core dependencies
gradio>=4.19.2
openai>=1.12.0
pandas>=2.2.0
numpy>=1.26.4
python-dotenv>=1.0.1
pyyaml>=6.0.1

# LlamaIndex and related packages
llama-index>=0.9.48
llama-index-llms-openai>=0.1.6
llama-index-embeddings-openai>=0.1.6
llama-parse>=0.1.2

# Machine Learning and Deep Learning
torch>=2.2.0
transformers>=4.37.2

# Document Processing
python-docx>=1.0.0
pypdf>=4.0.0
openpyxl>=3.1.2

# Image Processing
pillow>=10.2.0

# Vector Database
kdbai_client>=1.4.0

# Additional Utilities
scikit-learn>=1.4.0
tqdm>=4.66.1
aiohttp>=3.9.3
requests>=2.31.0

plotly